requirements.txt
Flask
Flask-SocketIO
Flask-Session
Pillow
numpy
opencv-python-headless
gunicorn
eventlet